02-github-agent/requirements.txt
PyGithub==2.1.1
python-dotenv==1.0.0
langchain-core==0.3.35
openai==1.58.1
langchain==0.3.18
langchain-anthropic==0.3.6
langchain-community==0.3.16
langchain-openai==0.3.6
streamlit==1.36.0